requirements.txt
pyOpenSSL==0.13.1
stomp.py==4.0.4
